requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
funcsigs==1.0.2
lark-parser==0.6.3
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
py==1.5.4
pytest==3.7.1
scandir==1.8
six==1.11.0
